src/common/types moved to src/types
