+++ src/app/user-profile/page.tsx
@@ -1,5 +1,5 @@
 import UserProfileCard from "@/components/profile/UserProfileCard";
-import { mockUser } from "@/common/mocks/user";
+import { mockUser } from "@/mocks/user";
 
 export default function UserProfilePage() {
   return <UserProfileCard user={mockUser} />;
